+++ Samples/iOS-Swift/iOS-Swift/AppDelegate.swift
@@ -23,6 +23,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         }
 
         SentrySDKWrapper.shared.startSentry()
+        
         SampleAppDebugMenu.shared.display()
         
         metricKit.receiveReports()